EM-2954 PathVariable changes         Updated test cases         moved one class temporarily to exclude the not implemeted test case         Deleted the not required dto and its implemetations
