+++ build.gradle
@@ -157,6 +157,7 @@ def coverageExclusionList = [
     '**uk/gov/hmcts/reform/dg/docassembly/config/logging/*',
     '**uk/gov/hmcts/reform/dg/docassembly/domain/*',
     '**uk/gov/hmcts/reform/dg/docassembly/service/dto/external/annotation/*',
+    '**uk/gov/hmcts/reform/dg/docassembly/service/impl/*',
     '**uk/gov/hmcts/reform/dg/docassembly/config/security/*'   
     ]
